Create and delete of msps events

+++ profile.php
@@ -109,7 +109,7 @@ else {
                   </td>
                   <td>
                     <a href="dist/uploads/<?php echo $row['unique_name'] ?>" download><i class="material-icons">cloud_download</i></a>
-                    <a href="delete-file.php?id=<?php echo $row['id'] ?>&name=<?php echo $row['unique_name']?>"><i class="material-icons">delete_forever</i></a>
+                    <a href="delete-file.php?id=<?php echo $row['id'] ?>&name=<?php echo $row['unique_name']?>" onclick="return confirm('¿Estás seguro de que deseas eliminar este archivo?')"><i class="material-icons">delete_forever</i></a>
                   </td>
                 </tr>
               <?php
@@ -122,7 +122,7 @@ else {
     <div id='test2' class='col s12'>
       <h3 class='center-align'>Eventos</h3>
       <div class="col s8 offset-s2">
-        <form class="form-event" id="formEvent" action="event.php" method="post">
+        <form class="form-event" id="formEvent" action="event.php" method="post" enctype='multipart/form-data'>
           <div class="row">
             <div class="input-field col s12">
                 <input id="title" type="text" class="validate" name="title">
@@ -131,10 +131,10 @@ else {
           </div>
           <div class="row">
             <div class="col s6">
-              <input type="text" name="date" value="" class="datepicker">
+              <input type="text" name="date" class="datepicker" placeholder="2018-12-31">
             </div>
             <div class="col s6">
-              <input type="text" name="hour" value="" class="timepicker">
+              <input type="text" name="hour" class="timepicker" placeholder="15:00">
             </div>
           </div>
           <div class="row">
@@ -151,7 +151,7 @@ else {
           </div>
           <div class="row">
             <div class=" input-field col s12">
-              <input type="file" name="image" class="validate">
+              <input type="file" name="file" class="validate">
               <label for="image">Imagen</label>
             </div>
           </div>
@@ -163,30 +163,49 @@ else {
           </div>
           <div class="row">
             <div class="input-field col s12 right-align">
-              <input type="submit" name="Enviar" value="Enviar" class="waves-effect waves-light btn yellow darken-1">
+              <input type="submit" name="submit" value="Enviar" class="waves-effect waves-light btn yellow darken-1">
             </div>
           </div>
         </form>
       </div>
       <div class="col s12">
-        <table>
-          <thead>
-            <tr>
+        <table class="responsive-table striped highlight">
+        <thead>
+          <tr>
               <th>Nombre</th>
               <th>Fecha</th>
               <th>Lugar</th>
               <th>Acciones</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr>
-              <td>1</td>
-              <td>1</td>
-              <td>1</td>
-              <td>1</td>
-            </tr>
-          </tbody>
-        </table>
+          </tr>
+        </thead>
+
+        <tbody>
+          <?php
+            $events = "SELECT * FROM event";
+            $result = $mysqli->query($events);
+
+            while ($row=$result->fetch_assoc()) {
+              ?>
+                <tr>
+                  <td>
+                    <?php echo $row['title'] ?>
+                  </td>
+                  <td>
+                    <?php echo $row['date_hour'] ?>
+                  </td>
+                  <td>
+                    <?php echo $row['place'] ?>
+                  </td>
+                  <td>
+                    <a href="edit-event.php?id=<?php echo $row['id'] ?>"><i class="material-icons">create</i></a>
+                    <a href="delete-event.php?id=<?php echo $row['id'] ?>&name=<?php echo $row['image'] ?>" onclick="return confirm('¿Estás seguro de que quieres borrar este evento?')"><i class="material-icons">delete_forever</i></a>
+                  </td>
+                </tr>
+              <?php
+            }
+          ?>
+        </tbody>
+      </table>
       </div>
     </div>
     <div id='test4' class='col s12'>
